+++ liquidation/src/main/java/cn/edu/xmu/other/liquidation/dao/LiquidationDao.java
@@ -99,5 +99,5 @@ public class LiquidationDao {
         }
     }
 
-    public ReturnObject getSimpleLiquInfo(Integer shopId,)
+    public ReturnObject getSimpleLiquInfo(Long shopId,)
 }
